remove thyme
