+++ cabal.project
@@ -1,6 +1,1 @@
 packages: .
-
-source-repository-package
-    type: git
-    location: https://github.com/kadena-io/thyme.git
-    tag: 5314711cf0e92c531cd49addcf9d9ffe466c2075
